RavenDB-15552 adjustments
