+++ src/Raven.Client/Json/Serialization/JsonDeserializationClient.cs
@@ -21,6 +21,7 @@ using Raven.Client.Documents.Operations.TimeSeries;
 using Raven.Client.Documents.Operations.TransactionsRecording;
 using Raven.Client.Documents.Queries;
 using Raven.Client.Documents.Queries.MoreLikeThis;
+using Raven.Client.Documents.Session.Operations;
 using Raven.Client.Documents.Session.TimeSeries;
 using Raven.Client.Documents.Smuggler;
 using Raven.Client.Documents.Subscriptions;
@@ -239,7 +240,7 @@ namespace Raven.Client.Json.Serialization
         internal static readonly ConcurrentDictionary<Type, Func<BlittableJsonReaderObject, TimeSeriesRangeResult>> CacheForTimeSeriesRangeResult =
             new ConcurrentDictionary<Type, Func<BlittableJsonReaderObject, TimeSeriesRangeResult>>();
 
-        internal static readonly Func<BlittableJsonReaderObject, GetRevisionsCountCommand.DocumentRevisionsCount> DocumentRevisionsCount = GenerateJsonDeserializationRoutine<GetRevisionsCountCommand.DocumentRevisionsCount>();
+        internal static readonly Func<BlittableJsonReaderObject, GetRevisionsCountOperation.DocumentRevisionsCount> DocumentRevisionsCount = GenerateJsonDeserializationRoutine<GetRevisionsCountOperation.DocumentRevisionsCount>();
 
 
     }
